refactored player movement
Player speed is calculated seperately now. `player.Climbing` is only
use in climbing code and not elsewhere (for collisions and climbing
speed); these places run a collision detection on the player with the
surrounding climable objects.

This has temporarily broken Climbing as the player's never colliding
with the climbable due to new collision detection but that will be
fixed shortly with hitbox offsets.