completely rewrote HasHitbox interface
Hitboxes are no-longer restricted to a single rectangle allowing for
more potential. Objects handle their own collision detection through
the `Overlaps` method allowing maths equations (like checking within
a circular range) to be used for collisions rather than the
`rectangle.Overlaps` which is hard set to a rectangle. `GetHibox`
returns a list of Rectangles. This is intended to be a simplified form
of the hitbox for another object to check against. `Origin` is pretty
self-explanitory, it's the origin of the object where it's position
is relative to.

Removed player river-jump (teleport) code temporarily due to new
complexities from new collision detection.