refactored lineart partial line bounds
the code for finding the start and end of the partial line segments
was very repetative. It was almost identical code for the start and
end just a few differences between Mins and Maxs. And then it's also
almost identical code for each side of the lineart (top, bottom...).

This has been split into a reuseable function `getPointOnBounds()`