implemented river priorities
when jumping a river it will now prioritise the shortest jump that can
be made rather than the first one found in the internal representation.
This makes the gameplay more predictable as it behaves based on things
the user can manipulate.

Also made river jumping work on the individual segments instead of only
the whole river.