implemented sunrise and sunset
In each day there is an hour long transition between day and night
and then night and day. This is a linear transition atm. If the sunrise
and sunset overlap due to it being towards one of the solstices then
the light will be set to the darkest of the 2.