changed states to only use strings
every value can be stored as a string. If it should be used in another
format then it can be converted to that when being read from the state
(possibly using helper methods like ints do currently).

This simplifies the code making it more readable and makes the future
json reading easier.