moved frame generation out of interface `Draw()` method
As `Draw()` was a method on the interface ebiten.Game it couldn't be
modified to return an error. So the part of the frame generation that
could cause an error has been moved to a separate method called in the
`Update()` method which does return an error. This is stored in a field
on the `Game` struct and then used in `Draw()` so that no errors need
to be used in there.