cached river and incline textures
rather than regenerating the textures and calculating the lineart on
every draw call, the textures are calculated initially and then when
something changes meaning a massive performance boost. When they need
to be recalculated the pointers to images in the cache can be set to nil
triggering a recalculation on the next draw call.