optimised river rendering
river rendering was causing a lot of lag in the game loop. This was due
to the lineart being generated every render cycle. It was meant to be
cached however development code was left in marking it dirty every time
it generates meaning that the cache was never used. Now that the
`markDirty` line has been removed profilers show that it uses much less
CPU power meaning it should fix github issue #3 where it caused lag on
lower power machines

TL;DR
river was't actually getting cached